+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 ### Removed
 
 - **BREAKING**: Remove default table style.
+- **BREAKING**: Remove the `TableView.hairlines` property.
 
 ## [0.15.0 - 2019-10-07](https://github.com/alonrbar/react-tablize/tree/v0.15.0)